requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow>=10.0.0
opencv-python==4.8.1.78
pytesseract>=0.3.10
numpy==1.26.4
pydantic>=2.0.0
python-jose[cryptography]==3.3.0

# YOLO and Face Recognition
ultralytics>=8.0.0
face-recognition>=1.3.0
dlib>=19.24.0
cmake>=3.25.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
requests==2.31.0
aiofiles>=0.8.0,<24.0.0
twilio>=7.0.0,<10.0.0
python-dotenv>=0.19.0
deepface>=0.0.79,<1.0.0
tensorflow>=2.16.0
tf-keras>=2.16.0

# MongoDB dependencies
motor>=2.5.0,<4.0.0
pymongo>=4.0.0,<5.0.0
dnspython>=2.0.0,<3.0.0

# HTTP client
httpx>=0.24.0,<1.0.0

# Redis for message queuing
redis==5.0.0

# Typing extensions for Python compatibility
typing_extensions>=4.6.0

# Fix for OpenSSL 1.0.2k compatibility
urllib3<2.0.0

# Note: TensorFlow and DeepFace are commented out due to compatibility issues
# Uncomment and test separately if needed:
# deepface>=0.0.75,<0.0.80
# tensorflow>=2.13.0,<2.16.0
# tf-keras>=2.13.0,<2.16.0
